requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
openai>=1.14.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
numpy>=1.26.0
Pillow>=10.2.0
pytesseract>=0.3.10
easyocr>=1.7.1
openai-whisper>=20231117
pydub>=0.25.1
scipy>=1.12.0
python-dotenv>=1.0.1
tiktoken>=0.6.0
plotly>=5.19.0
pandas>=2.2.0
sympy>=1.12
requests>=2.31.0
sounddevice>=0.4.6
soundfile>=0.12.1
